MyAgent/requirements.txt
langchain>=0.2.14
langchain-community>=0.2.10
langchain-core>=0.2.33
langchain-text-splitters>=0.2.2
langgraph>=0.2.14
langgraph-checkpoint-sqlite>=1.0.0
dashscope>=1.25.11
faiss-cpu>=1.13.2
python-dotenv>=1.2.1
httpx>=0.28.1
python-docx>=1.2.0
openpyxl>=3.1.5
python-pptx>=1.0.2
rank_bm25>=0.2.2
gradio>=4.0.0
ddgs>=9.0.0
Pillow>=10.0.0
